pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forge-memory"
version = "1.0.0"
description = "Persistent memory MCP server for AI agents"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
keywords = ["mcp", "memory", "ai", "agent", "sqlite", "persistence"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Database",
]
dependencies = [
    "mcp",
    "pyyaml",
]

[project.urls]
Homepage = "https://github.com/richardhapb/forge-memory"
Repository = "https://github.com/richardhapb/forge-memory"
Issues = "https://github.com/richardhapb/forge-memory/issues"

[project.scripts]
forge-memory = "forge_memory.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
]

[tool.coverage.run]
source = ["forge_memory"]
omit = ["forge_memory/__main__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 90
